Fix bug when syncing both ways
Previously, if used in push more or pull mode, the program worked just
fine, but if used in two-way mode, a bug would cause the program to
copy ALL files in the Linux directory to the Windows directory,
regardless of whether they were up-to-date or not.

This was because a list structure was used to track which files in which
directory neededto be transferred to the other, and files that didn't
were removed. The Linux directory is checked first, and if all files
were up to date with the Windows files, they are removed from the list
of Windows files to be transferred. Then the Windows directory is
checked, but because all the files were removed, the program thinks that
they're missing from the directory itself and need to be transferred
over, so it copies everything from the Linux directory to the Windows
directory.

This was fixed by changing the data structure containing the list of
files to be transferred from a simple list to a dictionary, with the
directory files as the key, and a boolean value as the associated value
(True for files to be transferred, False for files that don't need to be
transferred).